refactor: ChatMessage, ChatMessageMention 엔티티 패키지 변경
