+++ src/main/java/com/mobble/mobbleserver/domain/chat/message/MessageType.java
@@ -1,0 +1,5 @@
+package com.mobble.mobbleserver.domain.chat.message;
+
+public enum MessageType {
+    TEXT
+}
