+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/dto/response/ChatMessageResponseDto.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.chat.chatMessage.dto.response;
 
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.ChatMessage;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.MessageType;
+import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
+import com.mobble.mobbleserver.domain.chat.message.MessageType;
 import com.mobble.mobbleserver.util.DateTimeUtils;
 
 import java.time.LocalDateTime;
