--- src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/entity/ChatMessage.java
@@ -1,63 +1,0 @@
-package com.mobble.mobbleserver.refactor.chat.chatMessage.entity;
-
-import com.mobble.mobbleserver.common.baseEntity.CreatedAtEntity;
-import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
-import com.mobble.mobbleserver.domain.member.Member;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@Entity
-@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChatMessage extends CreatedAtEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "chat_message_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_room_id", nullable = false)
-    private ChatRoom chatRoom;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "sender_id", nullable = false)
-    private Member sender;
-
-    @Column(name = "content", nullable = false)
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "type", nullable = false)
-    private MessageType type;
-
-    @Builder(access = AccessLevel.PRIVATE)
-    private ChatMessage(
-            ChatRoom chatRoom,
-            Member sender,
-            String content,
-            MessageType type
-    ) {
-        this.chatRoom = chatRoom;
-        this.sender = sender;
-        this.content = content;
-        this.type = type;
-    }
-
-    public static ChatMessage createChatMessage(
-            ChatRoom chatRoom,
-            Member sender,
-            String content,
-            MessageType type
-    ) {
-        return ChatMessage.builder()
-                .chatRoom(chatRoom)
-                .sender(sender)
-                .content(content)
-                .type(type)
-                .build();
-    }
-}
