--- src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/entity/MessageType.java
@@ -1,5 +1,0 @@
-package com.mobble.mobbleserver.refactor.chat.chatMessage.entity;
-
-public enum MessageType {
-    TEXT
-}
