+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/repository/ChatMessageRepository.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.chatMessage.repository;
 
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.ChatMessage;
+import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long>, ChatMessageQueryRepository {
