+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/service/ChatMessageService.java
@@ -5,8 +5,8 @@ import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.dto.request.ChatMessageRequestDto;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.dto.response.ChatMessageResponseDto;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.ChatMessage;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.MessageType;
+import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
+import com.mobble.mobbleserver.domain.chat.message.MessageType;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.repository.ChatMessageRepository;
 import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import com.mobble.mobbleserver.refactor.chat.chatRoom.validator.ChatRoomValidator;
