+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/validator/ChatMessageValidator.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.chatMessage.validator;
 
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.ChatMessage;
+import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.repository.ChatMessageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
