+++ src/main/java/com/mobble/mobbleserver/refactor/chat/clubChatRoom/dto/request/ClubChatMessageRequestDto.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.clubChatRoom.dto.request;
 
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.MessageType;
+import com.mobble.mobbleserver.domain.chat.message.MessageType;
 
 public record ClubChatMessageRequestDto(
         Long chatRoomId,
