+++ src/main/java/com/mobble/mobbleserver/refactor/chat/directChatRoom/dto/request/DirectChatMessageRequestDto.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.directChatRoom.dto.request;
 
-import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.MessageType;
+import com.mobble.mobbleserver.domain.chat.message.MessageType;
 
 public record DirectChatMessageRequestDto(
         Long chatRoomId,
